LayoutTests/http/tests/permissions/permission-state-for-notifications-in-ephemeral-session.html

<!-- webkit-test-runner [ useEphemeralSession=true ] -->
<!DOCTYPE html>
<html>
<head>
<script src="/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This test checks that Permissions::query() is consistent with Notification.permission in ephemeral sessions");

jsTestIsAsync = true;

(async () => {
    notificationState = window?.Notification?.permission || "default";
    permissionsState = (await navigator.permissions.query({ name: "notifications" })).state;
    shouldBeEqualToString("notificationState", "default");
    shouldBeEqualToString("permissionsState", "prompt");
    finishJSTest();
})();
</script>
<script src="/resources/js-test-post.js"></script>
</body>
</html>
