LayoutTests/http/tests/permissions/permission-state-for-notifications-in-ephemeral-session-expected.txt

This test checks that Permissions::query() is consistent with Notification.permission in ephemeral sessions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS notificationState is "default"
PASS permissionsState is "prompt"
PASS successfullyParsed is true

TEST COMPLETE
